rpk: move cmd/cluster to its own package
